[Y26W2-136] refactor(ui): 컴포넌트 폴더 구조 변경
