+++ packages/ui/src/components/card/card.variant.ts
