+++ packages/ui/src/components/card/index.tsx
@@ -4,13 +4,13 @@ import IcInfo from "@/assets/icons/ic_info.svg?react";
 import IcLocation from "@/assets/icons/ic_location.svg?react";
 import IcMemo from "@/assets/icons/ic_memo.svg?react";
 import IcWalker from "@/assets/icons/ic_walker.svg?react";
-import { Button } from "@/button";
-import { Chip } from "@/chip";
-import Divider from "@/divider/divider";
-import { AddIconButton } from "@/icon-button/add-icon-button";
-import { DeleteIconButton } from "@/icon-button/delete-icon-button";
-import IconText from "@/icon-text";
-import { ImageCard } from "@/image-card";
+import { Button } from "@/components/button";
+import { Chip } from "@/components/chip";
+import Divider from "@/components/divider/divider";
+import { AddIconButton } from "@/components/icon-button/add-icon-button";
+import { DeleteIconButton } from "@/components/icon-button/delete-icon-button";
+import IconText from "@/components/icon-text";
+import { ImageCard } from "@/components/image-card";
 import { cn } from "@/utils";
 import { chip } from "./card.variant";
