+++ packages/ui/src/components/icon-button/delete-icon-button/delete-icon-button.stories.tsx
