+++ packages/ui/src/components/icon-button/floating-icon-button/floating-icon-button.stories.tsx
