+++ packages/ui/src/components/icon-button/solid-expand/solid-expand.stories.tsx
