+++ packages/ui/src/components/icon-tabs/index.stories.tsx
