+++ packages/ui/src/index.ts
@@ -1,19 +1,19 @@
 export { default as IcList } from "@/assets/icons/ic_list.svg?react";
 export { default as IcTable } from "@/assets/icons/ic_table.svg?react";
-export { Button } from "@/button";
-export { Chip } from "@/chip";
-export { Graph } from "@/graph";
-export { AddIconButton } from "@/icon-button/add-icon-button";
-export { CheckIconButton } from "@/icon-button/check-icon-button";
-export { DeleteIconButton } from "@/icon-button/delete-icon-button";
-export { FloatingIconButton } from "@/icon-button/floating-icon-button";
-export { IconButton } from "@/icon-button/normal-icon-button";
-export { SolidExpand } from "@/icon-button/solid-expand";
-export type { IconTabsOption, IconTabsProps } from "@/icon-tabs";
-export { IconTabs } from "@/icon-tabs";
-export { InterimButton } from "@/interim-button";
-export { InterimDescription } from "@/interim-description";
-export { InterimTitle } from "@/interim-title";
-export { Pin } from "@/pin";
-export { TextField } from "@/text-field";
+export { Button } from "@/components/button";
+export { Chip } from "@/components/chip";
+export { Graph } from "@/components/graph";
+export { AddIconButton } from "@/components/icon-button/add-icon-button";
+export { CheckIconButton } from "@/components/icon-button/check-icon-button";
+export { DeleteIconButton } from "@/components/icon-button/delete-icon-button";
+export { FloatingIconButton } from "@/components/icon-button/floating-icon-button";
+export { IconButton } from "@/components/icon-button/normal-icon-button";
+export { SolidExpand } from "@/components/icon-button/solid-expand";
+export type { IconTabsOption, IconTabsProps } from "@/components/icon-tabs";
+export { IconTabs } from "@/components/icon-tabs";
+export { InterimButton } from "@/components/interim-button";
+export { InterimDescription } from "@/components/interim-description";
+export { InterimTitle } from "@/components/interim-title";
+export { Pin } from "@/components/pin";
+export { TextField } from "@/components/text-field";
 export { cn } from "@/utils";
